requirements.txt
psycopg2
pprint
schedule
python-dateutil
xmltodict
requests
bottlenose
schedule
bs4
argparse
google-cloud-storage